# backend/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Silence Notes API
  description: |
    ## Overview

    Silence Notes is a Chrome extension for note-taking with hashtag filtering and persistent backend storage.
    This API provides authentication, user management, and note functionality with comprehensive security features.

    ## Authentication

    The API uses Google OAuth 2.0 for authentication with JWT tokens. Protected endpoints require a valid Bearer token:

    ```
    Authorization: Bearer <your-jwt-token>
    ```

    ## Security Features

    - **Rate Limiting**: Global (100 req/sec) and per-user (60 req/min) limits
    - **CORS Protection**: Configured for Chrome extension and web origins
    - **Security Headers**: CSP, HSTS, XSS protection, frame options
    - **Session Management**: Concurrent session limits and activity tracking
    - **Request Validation**: Input sanitization and size limits

    ## Error Handling

    All errors return JSON with consistent format:
    ```json
    {
      "error": "Error message",
      "code": 400,
      "success": false
    }
    ```

    ## Rate Limiting

    Rate limit headers are included in responses:
    - `X-RateLimit-Limit`: Maximum requests per window
    - `X-RateLimit-Remaining`: Remaining requests in current window
    - `X-RateLimit-Reset`: Time when rate limit window resets

  version: 1.0.0
  contact:
    name: Silence Notes Team
    email: support@silenzenotes.com
  license:
    name: MIT
...
servers:
...
paths:
...
  /user/profile:
    get:
      tags:
        - User
      summary: Get user profile
      description: Retrieves the current user's profile information including preferences.
      operationId: getUserProfile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
              example:
                success: true
                user:
                  id: "123e4567-e89b-12d3-a456-426614174000"
                  google_id: "123456789012345678901"
                  email: "user@example.com"
                  name: "John Doe"
                  avatar_url: "https://lh3.googleusercontent.com/..."
                  created_at: "2024-01-01T10:00:00Z"
                  updated_at: "2024-01-01T11:00:00Z"
                  preferences:
                    theme: "light"
                    language: "en"
                    timezone: "UTC"
                    email_notifications: true
                    auto_save: true
                    default_note_view: "grid"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "User not found"
                code: 404
                success: false
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - User
      summary: Update user profile
      description: Updates the current user's profile information.
      operationId: updateUserProfile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfileUpdateRequest'
            example:
              name: "John Smith"
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
              example:
                success: true
                message: "Profile updated successfully"
                user:
                  id: "123e4567-e89b-12d3-a456-426614174000"
                  name: "John Smith"
                  email: "user@example.com"
                  updated_at: "2024-01-01T12:00:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /user/preferences:
    get:
      tags:
        - User
      summary: Get user preferences
      description: Retrieves the current user's preferences.
      operationId: getUserPreferences
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User preferences retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPreferencesResponse'
              example:
                success: true
                preferences:
                  theme: "light"
                  language: "en"
                  timezone: "UTC"
                  email_notifications: true
                  auto_save: true
                  default_note_view: "grid"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - User
      summary: Update user preferences
      description: Updates the current user's preferences.
      operationId: updateUserPreferences
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPreferencesUpdateRequest'
            example:
              theme: "dark"
              language: "fr"
              email_notifications: false
      responses:
        '200':
          description: Preferences updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPreferencesResponse'
              example:
                success: true
                message: "Preferences updated successfully"
                preferences:
                  theme: "dark"
                  language: "fr"
                  timezone: "UTC"
                  email_notifications: false
                  auto_save: true
                  default_note_view: "grid"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /user/sessions:
    get:
      tags:
        - User
      summary: Get user sessions
      description: Retrieves all active sessions for the current user.
      operationId: getUserSessions
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User sessions retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSessionsResponse'
              example:
                success: true
                sessions:
                  - id: "sess_123e4567-e89b-12d3-a456-426614174000"
                    ip_address: "192.168.1.100"
                    user_agent: "Mozilla/5.0 (Windows NT 10.0; Win64; x64)..."
                    created_at: "2024-01-01T10:00:00Z"
                    last_seen: "2024-01-01T11:30:00Z"
                    is_active: true
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /user/sessions/{sessionId}:
    delete:
      tags:
        - User
      summary: Delete user session
      description: Deletes a specific session for the current user.
      operationId: deleteUserSession
      security:
        - BearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          description: ID of the session to delete
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Session deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                success: true
                message: "Session deleted successfully"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Session not found"
                code: 404
                success: false
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT token obtained from authentication flow"
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique user identifier
          example: "123e4567-e89b-12d3-a456-426614174000"
        google_id:
          type: string
          description: Google account ID
          example: "123456789012345678901"
        email:
          type: string
          format: email
          description: User's email address
          example: "user@example.com"
        name:
          type: string
          description: User's display name
          example: "John Doe"
        avatar_url:
          type: string
          format: uri
          description: URL to user's profile picture
          example: "https://lh3.googleusercontent.com/..."
        created_at:
          type: string
          format: date-time
          description: When the user account was created
          example: "2024-01-01T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          description: When the user account was last updated
          example: "2024-01-01T11:00:00Z"
    UserPreferences:
      type: object
      properties:
        theme:
          type: string
          enum: [light, dark, auto]
          description: UI theme preference
          example: light
        language:
          type: string
          description: Interface language code
          example: en
        timezone:
          type: string
          description: User's timezone
          example: UTC
        email_notifications:
          type: boolean
          description: Whether email notifications are enabled
          example: true
        auto_save:
          type: boolean
          description: Whether notes are automatically saved
          example: true
        default_note_view:
          type: string
          enum: [grid, list]
          description: Default view for notes
          example: grid
    UserProfileResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        user:
          allOf:
            - $ref: '#/components/schemas/User'
            - type: object
              properties:
                preferences:
                  $ref: '#/components/schemas/UserPreferences'
    UserProfileUpdateRequest:
      type: object
      properties:
        name:
          type: string
          description: New display name for the user
          example: "John Smith"
          minLength: 1
          maxLength: 100
    UserPreferencesResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        preferences:
          $ref: '#/components/schemas/UserPreferences'
    UserPreferencesUpdateRequest:
      type: object
      properties:
        theme:
          type: string
          enum: [light, dark, auto]
          example: dark
        language:
          type: string
          example: fr
        timezone:
          type: string
          example: America/New_York
        email_notifications:
          type: boolean
          example: false
        auto_save:
          type: boolean
          example: true
        default_note_view:
          type: string
          enum: [grid, list]
          example: list
    UserSession:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Session identifier
          example: "sess_123e4567-e89b-12d3-a456-426614174000"
        ip_address:
          type: string
          description: IP address of the session
          example: "192.168.1.100"
        user_agent:
          type: string
          description: User agent string of the session
          example: "Mozilla/5.0 (Windows NT 10.0; Win64; x64)..."
        created_at:
          type: string
          format: date-time
          description: When the session was created
          example: "2024-01-01T10:00:00Z"
        last_seen:
          type: string
          format: date-time
          description: When the session was last active
          example: "2024-01-01T11:30:00Z"
        is_active:
          type: boolean
          description: Whether the session is currently active
          example: true
    UserSessionsResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        sessions:
          type: array
          items:
            $ref: '#/components/schemas/UserSession'
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          description: Success message
          example: "Operation completed successfully"
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Invalid request"
        code:
          type: integer
          description: HTTP status code
          example: 400
        success:
          type: boolean
          example: false
  responses:
    BadRequest:
      description: Bad request - invalid input data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Invalid request data"
            code: 400
            success: false
    Unauthorized:
      description: Unauthorized - authentication required or failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Authentication required"
            code: 401
            success: false
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Internal server error"
            code: 500
            success: false
...